Change Notes: - Refactored AudioProjectCompiler, encapsulating logic into classes and running it all from the AudioProjectService so it's consistent with saving and loading audio projects.
